refactor(ai-service): Step 10 — application pipelines
The use-case orchestrators move out of services/ and services/llm/
into app/pipelines/, matching the diagram's APPLICATION band.

Moves:
  app/services/ai_service.py
    → app/pipelines/chat/__init__.py     (AIService is the public entry)
  app/services/llm/telemetry_llm_orchestrator.py
    → app/pipelines/chat/orchestrator.py
  app/services/llm/annotation_pipeline.py
    → app/pipelines/annotation/__init__.py    (run_annotation is the public entry)
  app/services/llm/annotation/flows/*
    → app/pipelines/annotation/flows/*
  app/services/llm/annotation/agents/*
    → app/pipelines/annotation/agents/*
  app/services/llm/annotation/tools/*
    → app/pipelines/annotation/tools/*
  app/services/llm/annotation/{followup,results,formatters}.py
    → app/pipelines/annotation/{followup,results,formatters}.py
  app/services/llm/classifier_prompt_generation.py
    → app/ml/prompts.py     (per Page 4 migration map)

The whole app/services/llm/ subtree is now gone. The annotation
package's side-effect imports (formatters/agents/tools registration)
move from annotation/__init__.py to pipelines/annotation/__init__.py
so the public entry preserves the same registration-before-call
guarantee callers depend on.

Importers updated (17 files): ai_service, telemetry_llm_orchestrator,
annotation_pipeline, classifier_prompt_generation references across
API, UI, scripts, and remaining service files.

Two latent path bugs surfaced and fixed:
  • app/agents/backends/local_vlm.py used parents[5] to climb to the
    project root — that was correct for the original
    services/llm/agent/backends/ location, but Step 9 made it
    parents[3]. Fix exposed when the annotation-pipeline import chain
    pulled local_vlm in.
  • app/llm/local_llm.py used parents[3] for the same reason — Step 7
    moved it one level shallower. Fix: parents[2].
Both followed the same family as the zarr fix in Step 4 — caching
file-relative paths needs adjustment when modules relocate.

Validated:
  • app.main imports clean
  • all four agents register on import (annotation_root + describe_graphs
    + label_verifier + zoom)
  • DEFAULT_HF_CACHE = /app/models/huggingface_cache  (was //models/…)
  • VLM models_dir = /app/models
  • /health 200

After this commit, app/services/ contains only the 5 ML-named files
plus segment_updater and telemetry_segment_visualizer. Step 11
(the ML split per Page 5) is next.
