+++ acla_ai_service/ui/segment_tabs/components/manual_lap_agent_local.py
@@ -123,7 +123,7 @@ def render_lap_agent_local(df, session_id, selected_annotation_key, circuit_id, 
             key="lap_local_run", type="primary",
         ):
             try:
-                from app.services.llm.annotation_pipeline import (
+                from app.pipelines.annotation import (
                     AnnotationPipelineConfig,
                     run_annotation,
                 )
